Homework3rework
